+++ packages/ebsi/test/src/issuer_token_parameters/issuer_token_parameters_test.dart
@@ -1,13 +1,6 @@
-// Copyright (c) 2022, Very Good Ventures
-// https://verygood.ventures
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
 import 'package:flutter_test/flutter_test.dart';
 
-import '../test_class.dart';
+import 'issuer_token_parameters_class.dart';
 
 void main() {
   group('override test', () {
@@ -31,6 +24,11 @@ void main() {
       test(
         "algorithm is ES256 when key's curve is P-256",
         issuerTokenParameterTest.algorithmIsES256Test,
+      );
+
+      test(
+        'if alg is not null then return as it is',
+        issuerTokenParameterTest.algorithmIsNotNullTest,
       );
     });
